Source tracker and additional serialisers

+++ src/Kafka.Connect/Models/CommandContext.cs
@@ -7,7 +7,7 @@ public class CommandContext
     public string Connector { get; set; }
     public string Topic { get; set; }
     public int Partition { get; set; }
-    public int Offset { get; set; }
+    public long Offset { get; set; }
     public long Timestamp { get; set; } 
     public CommandConfig Command { get; set; }
 }
